requirements.txt
flask-restplus==0.13.0
Werkzeug==0.16.1
psycopg2-binary==2.8.5
gunicorn==20.0.4
Flask==1.1.2

# Packages for linting, testing and more
flake8==3.8.3
pytest==6.0.1
pytest-cov==2.10.1
requests==2.24.0